feat(app-services,gui): model library changes reach every open client
`AppEvent` has carried `ModelAdded`/`ModelRemoved`/`ModelUpdated` since it
existed, and `event_name()` has mapped all three — but nothing ever emitted
one. `ModelOps` mutated the library in silence, and the emitter the service
graph already owns was handed to `ServerOps` and withheld from `ModelOps`.

The frontend half was the same absence from the other side: `getEventCategory`
had no `model_` arm, so had the events been emitted, `validateEvent` would
have dropped every one without a word — the same silent-drop mechanism as the
MCP lane deleted in #851, except here the events are wanted.

This looked like nothing was wrong because each tab refetches after its own
edit, which is exactly the case that never needed an event. What did not work
was a second window, or a second browser tab, against the same daemon: one
client's edit was invisible to the other until someone hit refresh.

Scope, stated precisely: this covers clients attached to one running daemon.
It does **not** carry a `gglib model add` in a terminal through to an open
GUI — the CLI is a separate process, the broadcaster is an in-process
`tokio::broadcast`, and `gglib model add` does not route through `ModelOps` at
all. Downloads are already covered, by the download event stream.

So all three layers:

- `ModelOps` takes an emitter and announces add, update and remove. A refused
  remove stays silent; announcing a removal that did not happen would tell
  every client to drop a model still in the library.
- `ModelSummary` gains `From<&Model>`, so the five-field mapping is written
  once rather than at each emit site.
- The category router moves to `events/category.ts` — it is pure routing, and
  worth testing without standing up a connection — and gains the `model_` arm.
- `useModelLibraryEvents` subscribes and reloads, batching through the
  existing `createBatchWithinWindow` so a burst costs one refetch.

`ModelOps` construction in the CLI was spelled out three times, each with a
comment pointing at the last one; it is now `one_shot_model_ops`, where the
reasons for `NoopModelRuntime` and `NoopEmitter` live once.

Ratchets: models.rs 1110 → 516, with its tests moved to a `#[path]` sibling
following #860; models_tests.rs 766 recorded. sse.ts 353 → 342.
