+++ crates/gglib-cli/src/handlers/model/capabilities.rs
@@ -7,11 +7,8 @@
 //!
 //! [`ModelCapabilities`]: gglib_core::ModelCapabilities
 
-use std::sync::Arc;
-
 use anyhow::{Result, anyhow};
 use gglib_app_services::types::SetCapabilitiesRequest;
-use gglib_app_services::{ModelDeps, ModelOps};
 use gglib_core::ModelCapabilities;
 
 use super::resolver;
@@ -35,11 +32,7 @@ pub(crate) async fn execute(
     // `NoopModelRuntime` rather than `ctx.runner`: a one-shot CLI command has
     // no shared `ProcessManager` to check, and this handler never touches
     // serving status anyway (`get`/`set_capabilities` only).
-    let ops = ModelOps::new(ModelDeps {
-        core: ctx.app.clone(),
-        runtime: Arc::new(gglib_core::ports::NoopModelRuntime),
-        gguf_parser: ctx.gguf_parser.clone(),
-    });
+    let ops = super::one_shot_model_ops(ctx);
 
     // Read-only: no flags provided.
     if set.is_empty() && unset.is_empty() {
