+++ scripts/rust-complexity-baseline.txt
@@ -10,7 +10,8 @@ crates/gglib-app-services/src/benchmark/tune/scoring.rs 352
 crates/gglib-app-services/src/downloads.rs 375
 crates/gglib-app-services/src/launch_options.rs 327
 crates/gglib-app-services/src/mcp.rs 396
-crates/gglib-app-services/src/models.rs 1110
+crates/gglib-app-services/src/models_tests.rs 766
+crates/gglib-app-services/src/models.rs 516
 crates/gglib-app-services/src/proxy.rs 455
 crates/gglib-app-services/src/sampling_explain_tests.rs 694
 crates/gglib-app-services/src/sampling_explain.rs 477
