+++ scripts/ts-complexity-baseline.txt
@@ -16,7 +16,7 @@ src/pages/ChatPage.tsx 475
 src/pages/ModelControlCenterPage.tsx 318
 src/services/tools/registry.ts 432
 src/services/transport/api/client.ts 361
-src/services/transport/events/sse.ts 353
+src/services/transport/events/sse.ts 342
 src/services/transport/types/dashboard.ts 559
 src/styles/tailwind.css 402
 src/types/benchmark.ts 539
