+++ src/hooks/README.md
@@ -40,6 +40,7 @@ Custom React hooks for gglib GUI functionality.
 | Hook | Description |
 |------|-------------|
 | [`useModels.ts`](useModels.ts) | Model CRUD operations and listing |
+| [`useModelLibraryEvents.ts`](useModelLibraryEvents.ts) | Reload the library when another client changes it |
 | [`useServers.ts`](useServers.ts) | Server lifecycle management (start/stop/health) |
 | [`useTags.ts`](useTags.ts) | Model tagging operations |
 | [`useMcpServers.ts`](useMcpServers.ts) | MCP server configuration |
